refactor: useWebSocket 함수 사용할하도록 알림 기능 수정
